ws funcional

+++ src/main/java/dev/barrikeit/model/repository/RoleRepository.java
@@ -5,4 +5,4 @@ import dev.barrikeit.model.repository.base.GenericCodeRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends GenericCodeRepository<Role, Integer, String> {}
+public interface RoleRepository extends GenericCodeRepository<Role, Long, String> {}
